# docs/swagger.yaml
definitions:
...
  RestApi_UnUpset_internal_delivery_dto.CreateToDoRequest:
    properties:
      text:
        description: Текст задачи (обязательное поле, от 1 до 500 символов)
        maxLength: 500
        minLength: 1
        type: string
    required:
    - text
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.Response:
    properties:
      data:
        description: Данные ответа (опционально)
      message:
        description: Сообщение (обычно используется для ошибок)
        type: string
      status:
        description: Статус ответа (success/error)
        type: string
    type: object
...
  RestApi_UnUpset_internal_delivery_dto.ToDorResponse:
    properties:
      done:
        description: Статус выполнения задачи (true - выполнена, false - не выполнена)
        type: boolean
      id:
        description: Идентификатор задачи
        type: integer
      text:
        description: Текст задачи
        type: string
      user_id:
        description: Идентификатор пользователя, создавшего задачу
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для приложения UnUpset
  title: UnUpset API
  version: "1.0"
paths:
...
  /api/todos:
    get:
      description: Получение списка всех задач текущего пользователя (не выполненных)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ToDorResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение всех задач пользователя (не выполненных)
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Создание новой задачи для текущего пользователя
      parameters:
      - description: Данные задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.CreateToDoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ToDorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание новой задачи
      tags:
      - todos
  /api/todos/{id}:
    delete:
      description: Удаление существующей задачи
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Удаление задачи
      tags:
      - todos
    get:
      description: Получение конкретной задачи по её ID
      parameters:
      - description: ID задачи
        format: uint
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ToDorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение задачи по ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Обновление текста существующей задачи
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.CreateToDoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.ToDorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновление задачи
      tags:
      - todos
  /api/todos/{id}/done:
    patch:
      description: Отметить существующую задачу как выполненную
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/RestApi_UnUpset_internal_delivery_dto.Response'
      security:
      - ApiKeyAuth: []
      summary: Отметить задачу как выполненную
      tags:
      - todos
...
securityDefinitions:
  ApiKeyAuth:
    description: Для тестирования в Swagger UI (на макбуке логин работает только через
      хром, в сафари протестировать не получается)
    in: cookie
    name: mysession
    type: apiKey
swagger: "2.0"